fix(portal-web): fix proxy url when base path is not root
